+++ instant_xml_parser/lib/src/codegen/method_gen.dart
@@ -26,14 +26,15 @@ class MethodGenerator {
   final MethodInfo method;
   final SourceInfo symtable;
   final Iterable<CommonElement> commonElements;
+  final bool nullSafe;
   final Logger _log = Logger('MethodGenerator');
 
-  MethodGenerator(this.method, this.symtable)
+  MethodGenerator(this.method, this.symtable, this.nullSafe)
       : commonElements = method.commonElements(symtable);
 
   String get attributesBlock => [
         for (var f in commonElements.where((f) => f.field.isAttributeField))
-          AttributeFieldGenerator(f, method, symtable).toAction
+          AttributeFieldGenerator(f, method, symtable, nullSafe).toAction
       ].join('\n');
 
   Iterable<TagFieldGenerator> get childFieldGenerators => [
@@ -103,7 +104,7 @@ class MethodGenerator {
   String get resolveDeferredAttributes => [
         for (var f in commonElements
             .where((f) => f.field.isAttributeField && f.field.isDeferred))
-          AttributeFieldGenerator(f, method, symtable).resolveDeferred
+          AttributeFieldGenerator(f, method, symtable, nullSafe).resolveDeferred
       ].join('\n');
 
   String get startBlock => '''